Created a PlanBase class which you can insert Actions in. Everytime bwapi throws his update event, the planbase executes the actions. There are multiple insert options for the actions to insert:
insertFirst; inserts the action for specified units before all other actions of those units.
insertLast: inserts the aciton for specified units after all other actions of those units.
insertReplace: removes all actions for specified units and inserts the actions specified.
insertIdleOnly: inserts the actions for specified units only when their idle.

The planbase takes care that an action for a group of units is performed when all units are available.
So when e.g. specifying two single actions for unit u1, and one single action for unit 2. And then e.g. insertLast an action where both units are involved, unit2 will wait for unit1 to finish first his actions and than the group action is performed for both units at the same time. 


Action is an abstract class which has three important methods:
getUnitsInvolved() : Is callled by the plan base to determine for which units the action is.
isFinished() : When returned true, the action is removed from the planbase for units involved.
perform() : Is called each time bwapi throws an update event.